requirements.txt
streamlit>=1.31.0
anthropic>=0.18.1
aiohttp>=3.9.0
pandas>=2.1.4
python-dotenv>=1.0.0
supabase>=2.3.4

# Dash dependencies (for new CRM UI)
dash>=2.17.1
dash-mantine-components>=0.14.3
dash-iconify>=0.1.2
gunicorn>=21.2.0

# Multi-source web scraping (REQUIRES Python 3.10+)
# TODO: Uncomment after upgrading from Python 3.9.6 to 3.10+
# crawl4ai>=0.7.0
# beautifulsoup4>=4.12.0
# lxml>=4.9.0
